requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
crewai==0.16.3
langchain==0.1.9
astrapy==0.7.4
pydantic==2.6.1
python-multipart==0.0.9
jinja2==3.1.3
httpx==0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
numpy>=1.24.0
pandas>=2.1.0
tenacity>=8.2.0
aiohttp>=3.9.0
aiofiles>=0.8.0
sqlalchemy>=2.0.0
pinecone>=0.1.0
anthropic>=0.7.0
langchain-community>=0.0.10 